feat: add handle pending state of changeLanguage and (refactor: rename LanguageSeletionMenu) #1
