+++ packages/ui/src/types.ts
@@ -19,4 +19,5 @@ export enum ThemeMode {
 
 export enum Language {
   English = 'en',
+  Vietnamese = 'vi',
 }
